Rename customization folder to brand-customization
