+++ graylog2-web-interface/src/components/inputs/InputDiagnosis/DiagnosisMessageErrors.tsx
@@ -21,7 +21,7 @@ import { Section } from 'components/common';
 import StatusColorIndicator from 'components/common/StatusColorIndicator';
 import { ListGroup, ListGroupItem } from 'components/bootstrap';
 import usePluginEntities from 'hooks/usePluginEntities';
-import useProductName from 'customization/useProductName';
+import useProductName from 'brand-customization/useProductName';
 import HelpPopoverButton from 'components/common/HelpPopoverButton';
 
 import { DIAGNOSIS_HELP } from './Constants';
