+++ graylog2-web-interface/src/pages/IndexSetPage.tsx
@@ -39,7 +39,7 @@ import type { Indices } from 'stores/indices/IndicesStore';
 import { IndexerOverviewActions, IndexerOverviewStore } from 'stores/indexers/IndexerOverviewStore';
 import { IndexSetsActions, IndexSetsStore } from 'stores/indices/IndexSetsStore';
 import { IndicesActions, IndicesStore } from 'stores/indices/IndicesStore';
-import useProductName from 'customization/useProductName';
+import useProductName from 'brand-customization/useProductName';
 
 const REFRESH_INTERVAL = 2000;
