+++ graylog2-web-interface/src/pages/InputsPage.tsx
@@ -23,7 +23,7 @@ import useCurrentUser from 'hooks/useCurrentUser';
 import AppConfig from 'util/AppConfig';
 import { Link } from 'components/common/router';
 import Routes from 'routing/Routes';
-import useProductName from 'customization/useProductName';
+import useProductName from 'brand-customization/useProductName';
 
 const isCloud = AppConfig.isCloud();
