+++ graylog2-web-interface/src/pages/ServerUnavailablePage.tsx
@@ -25,7 +25,7 @@ import DocumentTitle from 'components/common/DocumentTitle';
 import { qualifyUrl } from 'util/URLUtils';
 import LoginChrome from 'components/login/LoginChrome';
 import type { ServerError } from 'stores/sessions/ServerAvailabilityStore';
-import useProductName from 'customization/useProductName';
+import useProductName from 'brand-customization/useProductName';
 
 const StyledIcon = styled(Icon)`
   margin-left: 6px;
