+++ graylog2-web-interface/src/pages/ShowMetricsPage.tsx
@@ -19,7 +19,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { ClusterNodeMetrics } from '@graylog/server-api';
 
-import useProductName from 'customization/useProductName';
+import useProductName from 'brand-customization/useProductName';
 import { DocumentTitle, PageHeader, Spinner } from 'components/common';
 import { MetricsComponent } from 'components/metrics';
 import type { Metric } from 'stores/metrics/MetricsStore';
